test(wasmplugin): simplify push and negative tests
